+++ tests/test_tensor/test_contract.py
@@ -1,8 +1,34 @@
-import pytest
 import numpy as np
+import pytest
 
 import quimb.tensor as qtn
 from quimb.tensor.contraction import _CONTRACT_BACKEND, _TENSOR_LINOP_BACKEND
+
+
+def test_tensor_contract_strip_exponent():
+    tn = qtn.TN_rand_reg(10, 3, 3, dtype=complex)
+    z0 = tn.contract()
+    m1, e1 = qtn.tensor_contract(*tn, strip_exponent=True)
+    assert m1 * 10**e1 == pytest.approx(z0)
+    # test tn.exponent is reinserted
+    tn.equalize_norms_(value=1.0)
+    z2 = tn.contract()
+    assert z2 == pytest.approx(z0)
+    # test tn.exponent is reinserted with strip exponent
+    m3, e3 = tn.contract(strip_exponent=True)
+    assert m3 * 10**e3 == pytest.approx(z0)
+    # test tn.exponent is not reinserted when specified
+    z4 = tn.contract(exponent=False)
+    assert z4 != pytest.approx(z0)
+    # test tn.exponent is not reinserted when specified with strip exponent
+    m5, e5 = tn.contract(strip_exponent=True, exponent=False)
+    assert m5 * 10**e5 != pytest.approx(z0)
+    # test explicit exponent
+    z6 = tn.contract(exponent=tn.exponent)
+    assert z6 == pytest.approx(z0)
+    # test explicit exponent with strip exponent
+    m7, e7 = tn.contract(strip_exponent=True, exponent=tn.exponent)
+    assert m7 * 10**e7 == pytest.approx(z0)
 
 
 class TestContractOpts: